# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "reason_code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SimulatorMutationResponse": {
        "additionalProperties": false,
        "properties": {
          "baseline_ticks_emitted": {
            "minimum": 0.0,
            "title": "Baseline Ticks Emitted",
            "type": "integer"
          },
          "baseline_ticks_required": {
            "minimum": 0.0,
            "title": "Baseline Ticks Required",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "metric_interval_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Metric Interval Seconds",
            "type": "integer"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "scenario_state": {
            "minLength": 1,
            "title": "Scenario State",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "source_health": {
            "items": {
              "$ref": "#/components/schemas/SourceHealth"
            },
            "title": "Source Health",
            "type": "array"
          },
          "sources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceCounters"
            },
            "title": "Sources",
            "type": "object"
          },
          "state": {
            "enum": [
              "stopped",
              "running",
              "paused",
              "ready",
              "triggering",
              "completed"
            ],
            "title": "State",
            "type": "string"
          },
          "virtual_clock": {
            "format": "date-time",
            "title": "Virtual Clock",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "state",
          "scenario_state",
          "scenario_id",
          "virtual_clock",
          "seed",
          "metric_interval_seconds",
          "baseline_ticks_emitted",
          "baseline_ticks_required",
          "sources",
          "source_health",
          "request_id"
        ],
        "title": "SimulatorMutationResponse",
        "type": "object"
      },
      "SimulatorResetResponse": {
        "additionalProperties": false,
        "properties": {
          "baseline_ticks_emitted": {
            "minimum": 0.0,
            "title": "Baseline Ticks Emitted",
            "type": "integer"
          },
          "baseline_ticks_required": {
            "minimum": 0.0,
            "title": "Baseline Ticks Required",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "metric_interval_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Metric Interval Seconds",
            "type": "integer"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "reset_audit_id": {
            "minLength": 1,
            "title": "Reset Audit Id",
            "type": "string"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "scenario_state": {
            "minLength": 1,
            "title": "Scenario State",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "source_health": {
            "items": {
              "$ref": "#/components/schemas/SourceHealth"
            },
            "title": "Source Health",
            "type": "array"
          },
          "sources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceCounters"
            },
            "title": "Sources",
            "type": "object"
          },
          "state": {
            "enum": [
              "stopped",
              "running",
              "paused",
              "ready",
              "triggering",
              "completed"
            ],
            "title": "State",
            "type": "string"
          },
          "virtual_clock": {
            "format": "date-time",
            "title": "Virtual Clock",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "state",
          "scenario_state",
          "scenario_id",
          "virtual_clock",
          "seed",
          "metric_interval_seconds",
          "baseline_ticks_emitted",
          "baseline_ticks_required",
          "sources",
          "source_health",
          "request_id",
          "reset_audit_id"
        ],
        "title": "SimulatorResetResponse",
        "type": "object"
      },
      "SimulatorStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "baseline_ticks_emitted": {
            "minimum": 0.0,
            "title": "Baseline Ticks Emitted",
            "type": "integer"
          },
          "baseline_ticks_required": {
            "minimum": 0.0,
            "title": "Baseline Ticks Required",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "metric_interval_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Metric Interval Seconds",
            "type": "integer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "scenario_state": {
            "minLength": 1,
            "title": "Scenario State",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "source_health": {
            "items": {
              "$ref": "#/components/schemas/SourceHealth"
            },
            "title": "Source Health",
            "type": "array"
          },
          "sources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceCounters"
            },
            "title": "Sources",
            "type": "object"
          },
          "state": {
            "enum": [
              "stopped",
              "running",
              "paused",
              "ready",
              "triggering",
              "completed"
            ],
            "title": "State",
            "type": "string"
          },
          "virtual_clock": {
            "format": "date-time",
            "title": "Virtual Clock",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "state",
          "scenario_state",
          "scenario_id",
          "virtual_clock",
          "seed",
          "metric_interval_seconds",
          "baseline_ticks_emitted",
          "baseline_ticks_required",
          "sources",
          "source_health"
        ],
        "title": "SimulatorStatusResponse",
        "type": "object"
      },
      "SourceCounters": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "minimum": 0.0,
            "title": "Accepted",
            "type": "integer"
          },
          "collapsed": {
            "minimum": 0.0,
            "title": "Collapsed",
            "type": "integer"
          },
          "emitted": {
            "minimum": 0.0,
            "title": "Emitted",
            "type": "integer"
          },
          "quarantined": {
            "minimum": 0.0,
            "title": "Quarantined",
            "type": "integer"
          }
        },
        "required": [
          "emitted",
          "accepted",
          "collapsed",
          "quarantined"
        ],
        "title": "SourceCounters",
        "type": "object"
      },
      "SourceHealth": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "minimum": 0.0,
            "title": "Accepted",
            "type": "integer"
          },
          "collapsed": {
            "minimum": 0.0,
            "title": "Collapsed",
            "type": "integer"
          },
          "fixture_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixture Version"
          },
          "last_ingest_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ingest At"
          },
          "quarantined": {
            "minimum": 0.0,
            "title": "Quarantined",
            "type": "integer"
          },
          "source_id": {
            "minLength": 1,
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "metrics",
              "logs",
              "alerts",
              "config_changes",
              "topology"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "enum": [
              "ready",
              "error"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "source_type",
          "status",
          "last_ingest_at",
          "accepted",
          "collapsed",
          "quarantined"
        ],
        "title": "SourceHealth",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Milestone-0 contract API; feature routes are owner-labelled stubs.",
    "title": "Network Anomaly Root-Cause Assistant",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/simulator/reset": {
      "post": {
        "description": "Execute the full demo reset sequence (blueprint \u00a75.2, P1-22).\n\nStops the simulator, clears demo rows, reloads topology,\nre-seeds history, resets the simulator clock, and writes DEMO_RESET audit.",
        "operationId": "reset_api_v1_simulator_reset_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorResetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Reset",
        "tags": [
          "simulator"
        ]
      }
    },
    "/api/v1/simulator/scenarios/{scenario_id}/trigger": {
      "post": {
        "operationId": "trigger_api_v1_simulator_scenarios__scenario_id__trigger_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Trigger",
        "tags": [
          "simulator"
        ]
      }
    },
    "/api/v1/simulator/start": {
      "post": {
        "operationId": "start_api_v1_simulator_start_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start",
        "tags": [
          "simulator"
        ]
      }
    },
    "/api/v1/simulator/status": {
      "get": {
        "operationId": "status_api_v1_simulator_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Status",
        "tags": [
          "simulator"
        ]
      }
    },
    "/api/v1/simulator/stop": {
      "post": {
        "operationId": "stop_api_v1_simulator_stop_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Stop",
        "tags": [
          "simulator"
        ]
      }
    },
...
  }
}